requirements.txt
fastapi==0.123.5
uvicorn==0.38.0
jinja2==3.1.6
python-multipart==0.0.20
python-dotenv==1.2.1

pdfplumber==0.11.8
python-docx==1.2.0
python-pptx==1.0.2

sentence-transformers==5.1.2
faiss-cpu==1.13.0
numpy==2.2.6
scikit-learn==1.7.2
torch==2.9.1
transformers==4.57.3

groq==0.37.0
requests==2.32.5
